requirements.txt
apache-airflow
pandas
jupyter
matplotlib
requests
ipython
plotly
panel
kaleido
